Ensure that all distance calculations are performed in the Float_ type.
The Matrix_::data_type is only to be used for data storage; after an
observation is extracted, its coordinates are cast to Float_ for all
calculations and internal storage. This ensures that our internal
precision is consistent, i.e., we get the same results when a different
Matrix_::data_type is used (assuming the actual data values aren't
truncated by switching between data_types, of course).