Replace PCA-partitioning initialization with the simpler variance-partitioning. (#9)
Var-Part is much simpler and faster than PCA-part, while arguably being "good
enough" for an initial partitioning. It is also truly deterministic; in
contrast, PCA-part needs to use approximate PCA to speed things up, which
introduces randomization that weakens its deterministic claims.

To avoid an outright breaking change, the InitializePcaPartition class is now
an alias of InitializeVariancePartition. The two classes are pretty close in
concept and indeed, the latter is just a specialization of the former for
diagonal covariance matrices. The results should be pretty similar.

This change allows us to get rid of the powerit dependency. If we were to ever
re-introduce PCA-part, we should probably put it in an extension library to
reduce the number of transitive dependencies on the main kmeans library.