Optimizing each VarPart step to minimize the residual sum of squares.
We effectively do an exact 1-dimensional k-means (where k=2) along the chosen
dimension to decide each partition. This requires a little bit of work but
seems to yield more sensible partitions than just choosing the mean - and at
the very least, it's not worse - so we use it as our new default approach.