Do not quit Hartigan-Wong prematurely when quick_transfer does not converge. (#10)
This gives the opportunity for the algorithm to overcome the convergence
failure by performing more optimal transfers and trying another transfer - or
at least, just getting a few more steps in before we run out of iterations.

Users can restore the previous behavior (equivalent to R's stats::kmeans) with
a boolean switch in the Options. We also add another option to modify the
number of quick_transfer iterations, mostly for testing purposes.