Recompute centroids, WCSS loss for every outer Hartigan-Wong iteration.
This spends a little bit of compute to avoid accumulation of errors from
repeated transfers. For example, if a centroid moves so much that its
original set of assigned points is replaced, the new centroid would
contain the numerical error from subtracting the original centroid.

The errors accumulate in the centroids but the losses are calculated
from the centroids so we need to recompute them as well. This simplifies
the code a little bit as we no longer have to do a pass through the
dataset to recompute losses when quick_transfer fails to converge.