Overhaul the template parametrizations, add an actual Matrix interface class. (#13)
We explicitly list out the observation index and input data types in the
Initialize and Refine interface classes, as this is simpler to understand than
to fishing out the types from the Matrix_ class. It's easier for users to
decide on suitable basic types compared to figuring out what Matrix_ to use.

Making it easy for users to fill in the template parameters is important,
because we removed many of the previous default types. This favors the explicit
over the implicit and forces the user to be aware of the types they are using.
Which they must already know, because they'll need to supply the inputs (of the
specified type) and work with the outputs (of the specified type).

We also refactor MockMatrix into an actual Matrix interface class from which
SimpleMatrix derives. This allows our various functions/methods to default to
accepting a Matrix subclass, which allows run-time extensions to different
representations. Users can override this by forcing the use of a SimpleMatrix
in the template, which gives the compiler an opportunity for dervirtualization.