Added utility to remove unused centers from the k-means results.
It is hard for each Refine method to guarantee that avoid empty clusters will
not be generated, so we just provide a function that allows users to sanitize
the outputs afterwards. This is optional as some applications might not care
about empty clusters, in which case you can just use the Refine output as-is.

This function just moves empty clusters to the right of the centers matrix so
that they can be easily ignored, and we remap the cluster assignments to match.